fix previous sibling, w/ test
`beginning-of-line-text` ignores dashes, so in a yaml list starting with
the dash, the point didn't move, and thus the function thought the
previous sibling was itself.

We register the current indentation and then go to the (real) beginning
of line, to search for the sibling, now without danger of misleading.